backend/requirements.txt
Django==5.0.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
python-decouple==3.8
psycopg2-binary==2.9.9
Pillow==10.0.1
celery==5.3.4
redis==5.0.1
django-filter==23.5
djangorestframework-simplejwt==5.5.1
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
python-dateutil==2.8.2
django-extensions==3.2.3
